fix(runtime): keep cloudflare agents payload modules small
